refactored and moved a lot of code around to separate files
